+++ sway-lsp/benches/lsp_benchmarks/requests.rs
@@ -1,3 +1,5 @@
+use std::path::PathBuf;
+
 use criterion::{black_box, criterion_group, Criterion};
 use lsp_types::{
     CodeLens, CompletionResponse, DocumentSymbolResponse, Position, Range,
@@ -79,17 +81,7 @@ fn benchmarks(c: &mut Criterion) {
     });
 
     c.bench_function("code_lens", |b| {
-        b.iter(|| {
-            let mut result = vec![];
-            session.runnables.iter().for_each(|item| {
-                let runnable = item.value();
-                result.push(CodeLens {
-                    range: runnable.range(),
-                    command: Some(runnable.command()),
-                    data: None,
-                });
-            });
-        })
+        b.iter(|| capabilities::code_lens::code_lens(&session, &uri.clone()))
     });
 
     c.bench_function("on_enter", |b| {
